pyproject.toml
[tool.poetry]
name = "aws-cdk-fastapi-lambda"
version = "0.0.1"
description = "Sample project to illustrate a real quick FastAPI deployment on Lambda Functions"
authors = ["Santiago Garcia Arango <san99tiago@gmail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "2.83.1"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.group.dev.dependencies]
aws-lambda-powertools = {extras = ["all"], version = "^2.16.2"}
boto3 = "^1.26.153"
poethepoet = "^0.20.0"
pytest = "^7.3.2"
pytest-mock = "^3.11.1"
coverage = "^7.2.7"
moto = "^4.1.11"
fastapi = {extras = ["all"], version = "^0.98.0"}
mangum = "^0.17.0"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = [
    "cdk",
    "src/lambdas",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
test-unit = "coverage run -m pytest tests/unit"
test-unit-lambdas = "coverage run -m pytest tests/unit/lambdas"
test-unit-cdk = "coverage run -m pytest tests/unit/cdk"
test-unit-html = "coverage html"

[tool.coverage.run]
branch = true
source = ["src", "cdk"]
omit = [
    "**/__init__.py"
]

[tool.coverage.report]
show_missing = false
